Move content from catalog lib
