+++ apps/backend-functions/src/data/types.ts
@@ -37,7 +37,7 @@ export {
   UserRole
 } from '@blockframes/permissions/types';
 export { PublicUser } from '@blockframes/auth/types';
-export { RequestDemoInformations } from '@blockframes/catalog/demo-request.model';
+export { RequestDemoInformations } from '@blockframes/utils/request-demo';
 export { MovieAnalytics, EventAnalytics, StoreStatus, StoreConfig } from '@blockframes/movie/movie/+state/movie.firestore';
 export { NotificationOptions, NotificationType, NotificationDocument } from '@blockframes/notification/types';
 export { App } from '@blockframes/utils/apps';
