+++ apps/catalog/catalog/src/app/landing/pages/landing-page/landing-page.component.ts
@@ -3,7 +3,7 @@ import { AngularFireFunctions } from '@angular/fire/functions';
 import { FormGroup } from '@angular/forms';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { RouterQuery } from '@datorama/akita-ng-router-store';
-import { RequestDemoInformations, createDemoRequestInformations } from '@blockframes/catalog/demo-request.model';
+import { RequestDemoInformations, createDemoRequestInformations } from '@blockframes/utils/request-demo';
 
 @Component({
   selector: 'catalog-landing-page',
