+++ apps/catalog/catalog/src/app/marketplace/movie/search/search.component.html
@@ -225,7 +225,7 @@
               </span>
             </mat-expansion-panel> -->
           </ng-container>
-          <catalog-avails-filter [availsForm]="availsForm"></catalog-avails-filter>
+          <distribution-avails-filter [availsForm]="availsForm"></distribution-avails-filter>
           <article fxLayoutAlign="center center" class="filter-buttons">
             <ng-container *ngIf="!availsForm.get('isActive').value ; else active">
               <button mat-stroked-button color="primary" (click)="applyAvailsFilter()">
