+++ apps/catalog/catalog/src/app/marketplace/movie/search/search.module.ts
@@ -3,7 +3,7 @@ import { MovieCardModule } from "@blockframes/ui/movie-card/movie-card.module";
 import { MovieDisplayListModule } from '@blockframes/movie/movie/components/display-list/display-list.module';
 import { TranslateSlugModule } from "@blockframes/utils/pipes/translate-slug.module";
 import { MovieFormGenresModule } from "@blockframes/movie/movie/form/main/genres/genres.module";
-import { AvailsFilterModule } from '@blockframes/catalog/components/avails-filter/avails-filter.module';
+import { AvailsFilterModule } from '@blockframes/movie/distribution-deals/components/avails-filter/avails-filter.module';
 
 // Pages
 import { MarketplaceSearchComponent } from './search.component';
