+++ apps/catalog/catalog/src/app/marketplace/title/view/view.module.ts
@@ -8,11 +8,6 @@ import { FlexLayoutModule } from '@angular/flex-layout';
 import { MarketplaceMovieViewComponent } from './view.component';
 
 // Custom Modules
-import {
-  CatalogReservedTerritoriesModule,
-  CatalogProductionInformationsModule,
-  CatalogPromotionalElementsModule
-} from '@blockframes/catalog';
 import { MovieDisplayAvailabilitiesModule } from '@blockframes/movie/movie/components/display-availabilities/display-availabilities.module';
 import { MovieDisplayProductionModule } from '@blockframes/movie/movie/components/display-production/display-production.module';
 import { MovieDisplayPrincipalInformationsModule } from '@blockframes/movie/movie/components/display-principal-informations/display-principal-informations.module';
@@ -68,9 +63,6 @@ const routes = [{
     MovieDisplayAssetsModule,
     MovieDisplayKeywordsModule,
     MovieDisplayVersionInfoModule,
-    CatalogReservedTerritoriesModule,
-    CatalogProductionInformationsModule,
-    CatalogPromotionalElementsModule,
     MovieDisplayFilmInfoCardModule,
     ImageReferenceModule,
     // Material
