+++ libs/catalog/src/index.ts
@@ -1,11 +1,2 @@
-export * from './lib/catalog.module';
-export * from './lib/components/reserved-territories/reserved-territoires.module';
-export * from './lib/components/production-informations/production-informations.module';
-export * from './lib/components/promotional-elements/promotional-elements.module';
-export * from './lib/components/avails-filter/avails-filter.module';
 export * from './lib/demo-request.model';
-export * from './lib/filter.util';
 
-// Form
-
-export * from './lib/form/search.form';
