--- libs/catalog/src/lib/catalog.module.spec.ts
@@ -1,14 +1,0 @@
-import { async, TestBed } from '@angular/core/testing';
-import { CatalogModule } from './catalog.module';
-
-describe('CatalogModule', () => {
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      imports: [CatalogModule]
-    }).compileComponents();
-  }));
-
-  it('should create', () => {
-    expect(CatalogModule).toBeDefined();
-  });
-});
