--- libs/catalog/src/lib/catalog.module.ts
@@ -1,9 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-
-@NgModule({
-  imports: [
-    CommonModule
-  ]
-})
-export class CatalogModule {}
