--- libs/catalog/src/lib/components/production-informations/production-informations.component.html
@@ -1,19 +1,0 @@
-<div fxLayout="column nowrap" fxLayoutAlign="start start" fxLayoutGap="30px">
-  <article fxLayout="column">
-    <p class="bf-dark-text">
-      Production {{stakeholders.length <= 1 ? 'company':'companies'}}
-    </p>
-    <div *ngFor="let company of stakeholders; let i = index" fxLayout="row nowrap">
-      {{company.displayName}}
-    </div>
-  </article>
-  <article fxLayout="column">
-    <p class="bf-dark-text">Sales agent</p>
-    <ng-container *ngIf="salesAgentDealAvatar; else placeholder">
-      <img [imgRef]="salesAgentDealAvatar">
-    </ng-container>
-    <ng-template #placeholder>
-      <b>-</b>
-    </ng-template>
-  </article>
-</div>
