--- libs/catalog/src/lib/components/production-informations/production-informations.component.scss
@@ -1,3 +1,0 @@
-img {
-    width: 80px;
-}
