--- libs/catalog/src/lib/components/promotional-elements/promotional-elements.component.html
@@ -1,22 +1,0 @@
-<article *ngFor="let info of elements" class="link" fxLayout="column" fxLayoutAlign="start start">
-    <a *ngIf="promotionalElements.promo_reel_link" (click)="promoReelOpened.emit()" mat-stroked-button color="accent" [href]="info.media.url" target="_blank">
-      <mat-icon>play_arrow</mat-icon>
-      <b>Watch promo reel</b>
-    </a>
-    <a *ngIf="promotionalElements.scenario" mat-stroked-button color="accent" [href]="info.media.url" target="_blank">
-      <mat-icon>play_arrow</mat-icon>
-      <b>Download scenario</b>
-    </a>
-    <a *ngIf="promotionalElements.screener_link" mat-button [href]="info.media.url" target="_blank">
-      <mat-icon>play_arrow</mat-icon>
-      <span>Watch screener</span>
-    </a>
-    <a *ngIf="promotionalElements.trailer_link" mat-button [href]="info.media.url" target="_blank">
-      <mat-icon>play_arrow</mat-icon>
-      <span>Watch trailer</span>
-    </a>
-    <a *ngIf="promotionalElements.teaser_link" mat-button [href]="info.media.url" target="_blank">
-      <mat-icon>play_arrow</mat-icon>
-      <span>Watch teaser</span>
-    </a>
-  </article>
