--- libs/catalog/src/lib/components/promotional-elements/promotional-elements.component.scss
@@ -1,6 +1,0 @@
-a {
-  margin-top: 8px;
-  span, b {
-    font-size: 20px;
-  }
-}
