--- libs/catalog/src/lib/components/reserved-territories/reserved-territoires.module.ts
@@ -1,11 +1,0 @@
-import { FlexLayoutModule } from '@angular/flex-layout';
-import { CommonModule } from '@angular/common';
-import { CatalogReservedTerritoriesComponent } from './reserved-territories.component';
-import { NgModule } from '@angular/core';
-
-@NgModule({
-  declarations: [CatalogReservedTerritoriesComponent],
-  imports: [CommonModule, FlexLayoutModule],
-  exports: [CatalogReservedTerritoriesComponent]
-})
-export class CatalogReservedTerritoriesModule {}
