+++ libs/movie/src/lib/distribution-deals/+state/distribution-deal.service.ts
@@ -16,7 +16,7 @@ import { combineLatest } from 'rxjs';
 import { ContractVersion } from '@blockframes/contract/version/+state/contract-version.model';
 import { DistributionDealQuery } from './distribution-deal.query';
 import { Movie } from '@blockframes/movie/movie/+state';
-import { AvailsSearch } from '@blockframes/catalog/form/search.form';
+import { AvailsSearch } from '../form/search.form';
 import { Model } from '@blockframes/utils/static-model/staticModels';
 import { getFilterMatchingDeals, getDealsInDateRange } from '../create/availabilities.util';
