+++ libs/utils/src/index.ts
@@ -1,5 +1,3 @@
-export * from './lib/firequery/types';
-export * from './lib/firequery/firequery';
 export * from './lib/utils.module';
 export * from './lib/sentry.module';
 export * from './lib/abstract-state-guard';
